fix(图表): 修复线面图开启缩略轴时，纵轴最小值不立即自动计算的问题
